+++ tests/Rule/Nette/data/presenterInject.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Efabrica\PHPStanRules\Tests\Rule\Nette\data;
+
+class Service
+{
+
+}
+
+class InjectPresenter
+{
+    /** @var Service @inject */
+    public Service $service;
+}