+++ tests/Rule/Nette/data/presenterInjectAttribute.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Efabrica\PHPStanRules\Tests\Rule\Nette\data;
+
+use Nette\DI\Attributes\Inject;
+
+class Service
+{
+
+}
+
+class InjectAttributePresenter
+{
+    #[Inject]
+    public Service $service;
+}